+++ flink-python/pyflink/fn_execution/utils/operation_utils.py
@@ -17,25 +17,12 @@
 ################################################################################
 import datetime
 from collections import Iterable
-from enum import Enum
 from functools import partial
 
 from typing import Any, Tuple, Dict, List
 
 from pyflink.common import Row
-from pyflink.common.serializer import VoidNamespaceSerializer
-from pyflink.datastream import RuntimeContext
-from pyflink.datastream.time_domain import TimeDomain
-from pyflink.fn_execution import flink_fn_execution_pb2, pickle
-from pyflink.fn_execution.datastream.keyed_process_function import \
-    InternalKeyedProcessFunctionOnTimerContext, InternalKeyedProcessFunctionContext
-from pyflink.fn_execution.datastream.window_operator import WindowOperator
-from pyflink.fn_execution.state_impl import RemoteKeyedStateBackend
-from pyflink.fn_execution.timerservice_impl import TimerServiceImpl, InternalTimerImpl, \
-    InternalTimerServiceImpl
-from pyflink.fn_execution.utils.input_handler import TwoInputRowWithTimerHandler, \
-    OneInputRowWithTimerHandler
-from pyflink.fn_execution.utils.output_factory import RowWithTimerOutputFactory
+from pyflink.fn_execution import pickle
 from pyflink.serializers import PickleSerializer
 from pyflink.table import functions
 from pyflink.table.udf import DelegationTableFunction, DelegatingScalarFunction, \
@@ -286,216 +273,3 @@ def load_aggregate_function(payload):
         return cls()
     else:
         return pickle.loads(payload)
-
-
-def extract_data_stream_stateless_function(udf_proto, runtime_context):
-    """
-    Extracts user-defined-function from the proto representation of a
-    :class:`Function`.
-
-    :param udf_proto: the proto representation of the Python :class:`Function`
-    :param runtime_context: the streaming runtime context
-    """
-    func_type = udf_proto.function_type
-    UserDefinedDataStreamFunction = flink_fn_execution_pb2.UserDefinedDataStreamFunction
-    func = None
-
-    user_defined_func = pickle.loads(udf_proto.payload)
-    if func_type == UserDefinedDataStreamFunction.MAP:
-        func = user_defined_func.map
-    elif func_type == UserDefinedDataStreamFunction.FLAT_MAP:
-        func = user_defined_func.flat_map
-    elif func_type == UserDefinedDataStreamFunction.CO_MAP:
-        co_map_func = user_defined_func
-
-        def wrapped_func(value):
-            # value in format of: [INPUT_FLAG, REAL_VALUE]
-            # INPUT_FLAG value of True for the left stream, while False for the right stream
-            return co_map_func.map1(value[1]) if value[0] else co_map_func.map2(value[2])
-        func = wrapped_func
-    elif func_type == UserDefinedDataStreamFunction.CO_FLAT_MAP:
-        co_flat_map_func = user_defined_func
-
-        def wrapped_func(value):
-            if value[0]:
-                yield from co_flat_map_func.flat_map1(value[1])
-            else:
-                yield from co_flat_map_func.flat_map2(value[2])
-        func = wrapped_func
-
-    elif func_type == UserDefinedDataStreamFunction.TIMESTAMP_ASSIGNER:
-        extract_timestamp = user_defined_func.extract_timestamp
-
-        def wrapped_func(value):
-            pre_timestamp = value[0]
-            real_data = value[1]
-            return extract_timestamp(real_data, pre_timestamp)
-        func = wrapped_func
-
-    def open_func():
-        if hasattr(user_defined_func, "open"):
-            user_defined_func.open(runtime_context)
-
-    def close_func():
-        if hasattr(user_defined_func, "close"):
-            user_defined_func.close()
-
-    return func, open_func, close_func
-
-
-def extract_process_function(user_defined_function_proto, ctx, runtime_context):
-    process_function = pickle.loads(user_defined_function_proto.payload)
-    process_element = process_function.process_element
-
-    def wrapped_process_function(value):
-        # VALUE[CURRENT_TIMESTAMP, CURRENT_WATERMARK, NORMAL_DATA]
-        ctx.set_timestamp(value[0])
-        ctx.timer_service().advance_watermark(value[1])
-        output_result = process_element(value[2], ctx)
-        return output_result
-
-    def open_func():
-        if hasattr(process_function, "open"):
-            process_function.open(runtime_context)
-
-    def close_func():
-        if hasattr(process_function, "close"):
-            process_function.close()
-
-    return wrapped_process_function, open_func, close_func
-
-
-def extract_keyed_stateful_function(user_defined_function_proto,
-                                    keyed_state_backend: RemoteKeyedStateBackend,
-                                    runtime_context: RuntimeContext):
-    func_type = user_defined_function_proto.function_type
-    UserDefinedDataStreamFunction = flink_fn_execution_pb2.UserDefinedDataStreamFunction
-    payload = pickle.loads(user_defined_function_proto.payload)
-    internal_timer_service = InternalTimerServiceImpl(keyed_state_backend)
-
-    def state_key_selector(normal_data):
-        return Row(normal_data[0])
-
-    def user_key_selector(normal_data):
-        return normal_data[0]
-
-    def input_selector(normal_data):
-        return normal_data[1]
-
-    if func_type == UserDefinedDataStreamFunction.KEYED_PROCESS or \
-            func_type == UserDefinedDataStreamFunction.KEYED_CO_PROCESS:
-        timer_service = TimerServiceImpl(internal_timer_service)
-        on_timer_ctx = InternalKeyedProcessFunctionOnTimerContext(timer_service)
-        ctx = InternalKeyedProcessFunctionContext(timer_service)
-        process_function = payload
-        output_factory = RowWithTimerOutputFactory(VoidNamespaceSerializer())
-
-        def open_func():
-            if hasattr(process_function, "open"):
-                process_function.open(runtime_context)
-
-        def close_func():
-            if hasattr(process_function, "close"):
-                process_function.close()
-
-        if func_type == UserDefinedDataStreamFunction.KEYED_PROCESS:
-
-            def process_element(normal_data, timestamp: int):
-                ctx.set_timestamp(timestamp)
-                user_current_key = user_key_selector(normal_data)
-                ctx.set_current_key(user_current_key)
-                return process_function.process_element(input_selector(normal_data), ctx)
-
-            def on_event_time(internal_timer: InternalTimerImpl):
-                timestamp = internal_timer.get_timestamp()
-                state_current_key = internal_timer.get_key()
-                user_current_key = user_key_selector(state_current_key)
-
-                on_timer_ctx.set_current_key(user_current_key)
-                on_timer_ctx.set_time_domain(TimeDomain.EVENT_TIME)
-
-                return process_function.on_timer(timestamp, on_timer_ctx)
-
-            def on_processing_time(internal_timer: InternalTimerImpl):
-                timestamp = internal_timer.get_timestamp()
-                state_current_key = internal_timer.get_key()
-                user_current_key = user_key_selector(state_current_key)
-
-                on_timer_ctx.set_current_key(user_current_key)
-                on_timer_ctx.set_time_domain(TimeDomain.PROCESSING_TIME)
-
-                return process_function.on_timer(timestamp, on_timer_ctx)
-
-            input_handler = OneInputRowWithTimerHandler(
-                internal_timer_service,
-                keyed_state_backend,
-                state_key_selector,
-                process_element,
-                on_event_time,
-                on_processing_time,
-                output_factory)
-
-            process_element_func = input_handler.accept
-        elif func_type == UserDefinedDataStreamFunction.KEYED_CO_PROCESS:
-            input_handler = TwoInputRowWithTimerHandler(
-                ctx,
-                on_timer_ctx,
-                timer_service,
-                keyed_state_backend,
-                process_function,
-                output_factory)
-
-            process_element_func = input_handler.accept
-        else:
-            raise Exception("Unsupported func_type: " + str(func_type))
-    elif func_type == UserDefinedDataStreamFunction.WINDOW:
-        window_operation_descriptor = payload
-        window_assigner = window_operation_descriptor.assigner
-        window_trigger = window_operation_descriptor.trigger
-        allowed_lateness = window_operation_descriptor.allowed_lateness
-        window_state_descriptor = window_operation_descriptor.window_state_descriptor
-        internal_window_function = window_operation_descriptor.internal_window_function
-        window_serializer = window_operation_descriptor.window_serializer
-        keyed_state_backend._namespace_coder_impl = window_serializer._get_coder()
-        window_operator = WindowOperator(
-            window_assigner,
-            keyed_state_backend,
-            user_key_selector,
-            window_state_descriptor,
-            internal_window_function,
-            window_trigger,
-            allowed_lateness)
-        output_factory = RowWithTimerOutputFactory(window_serializer)
-
-        def open_func():
-            window_operator.open(runtime_context, internal_timer_service)
-
-        def close_func():
-            window_operator.close()
-
-        input_handler = OneInputRowWithTimerHandler(
-            internal_timer_service,
-            keyed_state_backend,
-            state_key_selector,
-            lambda n, t: window_operator.process_element(input_selector(n), t),
-            window_operator.on_event_time,
-            window_operator.on_processing_time,
-            output_factory)
-
-        process_element_func = input_handler.accept
-    else:
-        raise Exception("Unsupported func_type: " + str(func_type))
-
-    return process_element_func, open_func, close_func
-
-
-"""
-All these Enum Classes MUST be in sync with
-org.apache.flink.streaming.api.utils.PythonOperatorUtils if there are any changes.
-"""
-
-
-class KeyedProcessFunctionInputFlag(Enum):
-    EVENT_TIME_TIMER = 0
-    PROC_TIME_TIMER = 1
-    NORMAL_DATA = 2
