+++ flink-python/src/main/java/org/apache/flink/streaming/api/operators/python/PythonProcessOperator.java
@@ -31,6 +31,8 @@ import org.apache.flink.streaming.api.watermark.Watermark;
 import org.apache.flink.streaming.runtime.streamrecord.StreamRecord;
 import org.apache.flink.types.Row;
 
+import static org.apache.flink.python.Constants.STATELESS_FUNCTION_URN;
+
 /**
  * {@link PythonProcessOperator} is responsible for launching beam runner which will start a python
  * harness to execute user defined python ProcessFunction.
@@ -40,8 +42,6 @@ public class PythonProcessOperator<IN, OUT>
         extends OneInputPythonFunctionOperator<IN, OUT, Row, OUT> {
 
     private static final long serialVersionUID = 1L;
-
-    private static final String PROCESS_FUNCTION_URN = "flink:transform:process_function:v1";
 
     /** Reusable row for normal data runner inputs. */
     private transient Row reusableInput;
@@ -102,6 +102,6 @@ public class PythonProcessOperator<IN, OUT>
 
     @Override
     public String getFunctionUrn() {
-        return PROCESS_FUNCTION_URN;
+        return STATELESS_FUNCTION_URN;
     }
 }
